chore: Split nodestore into functional areas
A nodestore and its related types evolved over time becoming more
complex. This change splits up the functionality into different areas to
make it easier to find specific parts of the nodestore logic.

Doc comments about the layout are extensive and are not repeated here.

Other than moving things around, some additional changes are:
 - `UpdateError` was dead code. It was used in early development of
   merkle's insert but is no longer used.
 - Accessors for various members of NodeStoreHeader instead of
   referencing them directly. This makes auditing their use a little
   easier, and improves scoping and readability.
